refact: 프로젝트 구조 수정(GameHistory 모듈)
