+++ gotcha-domain/src/main/java/gotcha_domain/gamehistory/Difficulty.java
@@ -1,6 +1,6 @@
 package gotcha_domain.gamehistory;
 
 public enum Difficulty {
-    EASY, NORMAL, HARD
+    BASIC,
+    ADVANCED
 }
-
