+++ gotcha-domain/src/main/java/gotcha_domain/gamehistory/GameType.java~
@@ -7,6 +7,16 @@ public enum GameType {
     TRICK_MYOMYO("묘묘를 속여라", 2, 2),
     LULU_ART_EXAM("루루의 미대입시", 1, 1);
 
+
+    public static GameType getGameTypeByName(String gameTypeName) {
+        try {
+            return GameType.valueOf(gameTypeName);
+        } catch (IllegalArgumentException e) {
+            return null;
+        }
+    }
+
+
     private final String description;
     private final int minPlayers;
     private final int maxPlayers;
