+++ gotcha-socket/build.gradle
@@ -53,6 +53,9 @@ dependencies {
 
     //JavaTimeModule
     implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310'
+
+    //JPA
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 }
 
 bootJar {
