+++ gotcha-socket/src/main/java/socket_server/domain/game/meta/GameMeta.java
@@ -2,9 +2,9 @@ package socket_server.domain.game.meta;
 
 import lombok.Builder;
 import lombok.Data;
-import socket_server.domain.game.enumType.Difficulty;
 import socket_server.domain.game.enumType.GameStatus;
-import socket_server.domain.game.enumType.GameType;
+import gotcha_domain.gamehistory.GameType;
+import gotcha_domain.gamehistory.Difficulty;
 import socket_server.domain.game.model.Game;
 
 import java.util.Map;
