+++ gotcha-socket/src/main/java/socket_server/gamehistory/dto/UserGameHistoryDetailRes.java
@@ -1,25 +1,30 @@
-package Gotcha.domain.gamehistory.dto;
+package socket_server.gamehistory.dto;
 
 import gotcha_domain.gamehistory.UserGameHistory;
 import lombok.Builder;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Builder
-public record UserGameHistorySummaryRes(
+public record UserGameHistoryDetailRes(
         Long gameId,
         String gameType,
         String difficulty,
         LocalDateTime playedAt,
-        int score
+        int score,
+        List<UserGameHistoryRes> rounds
 ) {
-    public static UserGameHistorySummaryRes from(UserGameHistory userGameHistory){
-        return UserGameHistorySummaryRes.builder()
+    public static UserGameHistoryDetailRes from(UserGameHistory userGameHistory) {
+        return UserGameHistoryDetailRes.builder()
                 .gameId(userGameHistory.getGameHistory().getId())
                 .gameType(String.valueOf(userGameHistory.getGameHistory().getGameType()))
                 .difficulty(String.valueOf(userGameHistory.getGameHistory().getDifficulty()))
                 .playedAt(userGameHistory.getGameHistory().getCreatedAt())
                 .score(userGameHistory.getGameHistory().getScore())
+                .rounds(userGameHistory.getGameHistory().getRoundHistories().stream()
+                        .map(UserGameHistoryRes::from)
+                        .toList())
                 .build();
     }
 }
