+++ gotcha-socket/src/main/java/socket_server/gamehistory/dto/UserGameHistorySummaryRes.java
@@ -1,30 +1,25 @@
-package Gotcha.domain.gamehistory.dto;
+package socket_server.gamehistory.dto;
 
 import gotcha_domain.gamehistory.UserGameHistory;
 import lombok.Builder;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Builder
-public record UserGameHistoryDetailRes(
+public record UserGameHistorySummaryRes(
         Long gameId,
         String gameType,
         String difficulty,
         LocalDateTime playedAt,
-        int score,
-        List<UserGameHistoryRes> rounds
+        int score
 ) {
-    public static UserGameHistoryDetailRes from(UserGameHistory userGameHistory) {
-        return UserGameHistoryDetailRes.builder()
+    public static UserGameHistorySummaryRes from(UserGameHistory userGameHistory){
+        return UserGameHistorySummaryRes.builder()
                 .gameId(userGameHistory.getGameHistory().getId())
                 .gameType(String.valueOf(userGameHistory.getGameHistory().getGameType()))
                 .difficulty(String.valueOf(userGameHistory.getGameHistory().getDifficulty()))
                 .playedAt(userGameHistory.getGameHistory().getCreatedAt())
                 .score(userGameHistory.getGameHistory().getScore())
-                .rounds(userGameHistory.getGameHistory().getRoundHistories().stream()
-                        .map(UserGameHistoryRes::from)
-                        .toList())
                 .build();
     }
 }
