+++ gotcha-socket/src/main/java/socket_server/gamehistory/repository/GameHistoryRepository.java
@@ -1,7 +1,8 @@
-package Gotcha.domain.gamehistory.repository;
+package socket_server.gamehistory.repository;
 
 import gotcha_domain.gamehistory.GameHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GameHistoryRepository extends JpaRepository<GameHistory,Long> {
+
 }
