+++ gotcha-socket/src/main/java/socket_server/gamehistory/service/GameHistoryService.java
@@ -1,9 +1,9 @@
-package Gotcha.domain.gamehistory.service;
+package socket_server.gamehistory.service;
 
-import Gotcha.domain.gamehistory.dto.UserGameHistoryDetailRes;
-import Gotcha.domain.gamehistory.dto.UserGameHistorySummaryRes;
-import Gotcha.domain.gamehistory.exception.GameHistoryExceptionCode;
-import Gotcha.domain.gamehistory.repository.UserGameHistoryRepository;
+import socket_server.gamehistory.dto.UserGameHistoryDetailRes;
+import socket_server.gamehistory.dto.UserGameHistorySummaryRes;
+import socket_server.gamehistory.exception.GameHistoryExceptionCode;
+import socket_server.gamehistory.repository.UserGameHistoryRepository;
 import gotcha_common.exception.CustomException;
 import gotcha_domain.gamehistory.UserGameHistory;
 import lombok.RequiredArgsConstructor;
@@ -30,9 +30,7 @@ public class GameHistoryService {
         return UserGameHistoryDetailRes.from(userGameHistory);
     }
 
-    public void handleGameEnd(UserGameHistory userGameHistory) {
 
 
-    }
 
 }
